requirements.txt
altair>=4.2
streamlit
pandas
google-cloud-firestore